Fixed a bug with how the tantrix subpaths were rejoined. The genetic solver will sometimes find a solution now.

+++ trunk/source/com/becker/puzzle/tantrix/model/fitting/PrimaryPathFitter.java
@@ -49,7 +49,7 @@ public class PrimaryPathFitter extends AbstractFitter {
             if (nbr != null) {
                 PathColor pathColor = placement.getPathColor(i);
 
-                if (pathColor == primaryColor && pathColor == nbr.getPathColor((byte)(i+3))) {
+                if (pathColor == primaryColor && (pathColor == nbr.getPathColor(i+3))) {
                     return true;
                 }
             }
